printer: store distances internally as microns (#25)
* printer: store distances internally as microns

Swap internal distance values to microns, converted back to mm for
gcode output and additional API output
Preserve existing API fields, adding micron output as secondary option
Swap all float values read from config files to 64bit, following
serde's default read scheme and preventing any inaccuracies added by
downcasting

+++ src/display.rs
@@ -5,12 +5,12 @@ use png::Decoder;
 pub struct Frame {
     pub file_name: String,
     pub buffer: Vec<u8>,
-    pub exposure_time: f32,
+    pub exposure_time: f64,
     pub bit_depth: u8,
 }
 
 impl Frame {
-    pub fn from_vec(name: String, exposure_time: f32, data: Vec<u8>) -> Frame {
+    pub fn from_vec(name: String, exposure_time: f64, data: Vec<u8>) -> Frame {
         let decoder = Decoder::new(data.as_slice());
 
         let mut png_reader = decoder.read_info().expect("Unable to read PNG metadata");
